[Update] Move src/lib/custom-error.js to src/custom-error.js
