+++ src/actions/article.js
@@ -3,7 +3,7 @@ import { article as articleSchema } from '../schemas/index'
 import { camelizeKeys } from 'humps'
 import { formatUrl } from '../utils/index'
 import { normalize } from 'normalizr'
-import { InternalServerError, NotFoundError } from '../lib/custom-error'
+import { InternalServerError, NotFoundError } from '../custom-error'
 import * as types from '../constants/action-types'
 import fetch from 'isomorphic-fetch'
 import qs from 'qs'
