+++ src/actions/author-articles.js
@@ -5,7 +5,7 @@ import * as CONSTANTS from '../constants/index'
 import { MAX_ARTICLES_PER_FETCH, NUMBER_OF_FIRST_RESPONSE_PAGE, RETURN_DELAY_TIME } from '../constants/author-page'
 import { arrayOf, normalize } from 'normalizr'
 
-import { InternalServerError } from '../lib/custom-error'
+import { InternalServerError } from '../custom-error'
 import { article as articleSchema } from '../schemas/index'
 import { camelizeKeys } from 'humps'
 import fetch from 'isomorphic-fetch'
